+++ packages/smarthr-ui/src/components/Table/TdCheckbox.tsx
@@ -18,7 +18,7 @@ const classNameGenerator = tv({
       'shr-flex shr-justify-center shr-px-1 shr-py-0.75',
       '[&:not(:has([disabled]))]:shr-cursor-pointer',
     ],
-    wrapper: 'shr-p-0',
+    wrapper: 'shr-w-min shr-p-0',
     checkbox: ['shr-leading-[0]', '[&>span]:shr-translate-y-[unset]'],
   },
 })
@@ -37,8 +37,7 @@ export const TdCheckbox = forwardRef<HTMLInputElement, Omit<CheckboxProps, keyof
 
     return (
       // Td に必要な属性やイベントは不要
-      // contentWidth={0} で td をテーブルの計算上最小幅にする
-      <Td contentWidth={0} vAlign={vAlign} fixed={fixed} className={classNames.wrapper}>
+      <Td vAlign={vAlign} fixed={fixed} className={classNames.wrapper}>
         <label className={classNames.inner}>
           <Checkbox {...rest} ref={ref} className={classNames.checkbox} />
           {children && <VisuallyHiddenText>{children}</VisuallyHiddenText>}